Divided tests , separated utility for testing process monitor from tests
