+++ src/UI/UITest/pch.h
@@ -3,5 +3,6 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 #include <gmock/gmock-matchers.h>
+#include <windows.h>
 
 #endif // PCH_H
